+++ extensions/duckduckgo-image-search/CHANGELOG.md
@@ -1,5 +1,10 @@
 # DuckDuckGo Image Search Changelog
 
+## [Configurable Primary Action] - {PR_MERGE_DATE}
+
+- Added a new preference to choose between Quick Look or Open in Browser as the primary action
+- Added `Open With` option in the context menu to open images with other supported applications
+
 ## [Save image] - 2025-10-09
 
 - Added functionality to save images.